services/qwen2vl_service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.24.0
python-multipart>=0.0.9
transformers>=4.44.2
accelerate>=0.33.0
bitsandbytes>=0.43.3
sentencepiece>=0.2.0
pillow>=10.0.0
torchvision==0.19.0
